bug: wrong cpu count report, and change to report available threads instead
Since a recent change at #3, the reported cpu count was wrong. And that's
because it changed to make actual use of my own created function `cpu_count()`.
Problem was it should not always add +1, because thats only needed for range().

While I was at it, I also changed it to report count of threads that is
available to the script (which is basically all).