revert back to -m cd instead -m auto at default
I am not convinced that automatic selection of format based on filesize works
good enough as a default. Some PSP games in example are smaller size and would
be recognized as CD, while DVD is recommended as far as I know. Also PS2 games
converted as DVD plain and simple does not work in the tested emulator by me.
But in all cases CD works at least. For for now I don't want to default to
auto.