bug: when script converted to AppImage, it would break working directory
When converted to AppImage the integrated AppRun will change the working directory of the
script, in order to find the integrated libraries and executables. This however breaks the
script and we need to set the working directory to what it was before the call of the AppImage.
This however also breaks the AppImage then, therefore we need to change it back as soon as
all files are set.

Alongside this change, make use of fullpath() function to make sure input and output paths are
resolved early.

And use sys.argv[0] instead __file__ when the script is compiled with Nuitka. Otherwise
the programs like chdman in the same directory would be ignored.